tests/requirements-dev.txt
pyyaml>=5.4
pytest>=6.0
